+++ tests/test_collect.py
@@ -103,10 +103,10 @@ def test_collect_same_task_different_ways(tmp_path, path_extension):
 @pytest.mark.parametrize(
     ("task_files", "pattern", "expected_collected_tasks"),
     [
-        (["example_task.py"], "'*_task.py'", 1),
-        (["tasks_example.py"], "'tasks_*'", 1),
-        (["example_tasks.py"], "'*_tasks.py'", 1),
-        (["task_module.py", "tasks_example.py"], "'tasks_*.py'", 1),
+        (["example_task.py"], "['*_task.py']", 1),
+        (["tasks_example.py"], "['tasks_*']", 1),
+        (["example_tasks.py"], "['*_tasks.py']", 1),
+        (["task_module.py", "tasks_example.py"], "['tasks_*.py']", 1),
         (["task_module.py", "tasks_example.py"], "['task_*.py', 'tasks_*.py']", 2),
     ],
 )
@@ -117,13 +117,23 @@ def test_collect_files_w_custom_file_name_pattern(
         f"[tool.pytask.ini_options]\ntask_files = {pattern}"
     )
 
-    for file in task_files:
-        tmp_path.joinpath(file).write_text("def task_example(): pass")
+    for file_ in task_files:
+        tmp_path.joinpath(file_).write_text("def task_example(): pass")
 
     session = build(paths=tmp_path)
 
     assert session.exit_code == ExitCode.OK
     assert len(session.tasks) == expected_collected_tasks
+
+
+def test_error_with_invalid_file_name_pattern(runner, tmp_path):
+    tmp_path.joinpath("pyproject.toml").write_text(
+        "[tool.pytask.ini_options]\ntask_files = 'asds'"
+    )
+
+    result = runner.invoke(cli, [tmp_path.as_posix()])
+    assert result.exit_code == ExitCode.CONFIGURATION_FAILED
+    assert "'task_files' must be a list of patterns." in result.output
 
 
 @pytest.mark.unit()